+++ src/base/b-slider/b-slider.ts
@@ -10,6 +10,19 @@ import symbolGenerator from 'core/symbol';
 import iData, { component, prop, field, system, hook, watch, wait, p } from 'super/i-data/i-data';
 export * from 'super/i-data/i-data';
 
+export interface SlideRect extends ClientRect {
+	offsetLeft: number;
+}
+
+/**
+ * -1 - Previous
+ * 0  - Not changed
+ * 1  - Next
+ */
+export type SlideDirection = -1 | 0 | 1;
+export type OptionProps = ((el: unknown, i: number) => Dictionary) | Dictionary;
+export type OptionsIterator<T = bSlider> = (options: unknown[], ctx: T) => unknown[];
+
 export const
 	$$ = symbolGenerator();
 
@@ -25,35 +38,8 @@ export const sliderModes = {
 	slider: true
 };
 
-export interface SlideRect extends ClientRect {
-	offsetLeft: number;
-}
-
-/**
- * -1 - Previous
- * 0  - Not changed
- * 1  - Next
- */
-export type SlideDirection = -1 | 0 | 1;
 export type AlignType = keyof typeof alignTypes;
 export type Mode = keyof typeof sliderModes;
-export type OptionProps = ((el: unknown) => Dictionary) | Dictionary;
-
-/**
- * Returns true if the specified value is in the range X > 0 && X <= 1
- * @param v
- */
-export function isBetweenZeroAndOne(v: number): boolean {
-	return v > 0 && v <= 1;
-}
-
-/**
- * Returns true if the specified value is greater than 0 and isn't infinite
- * @param v
- */
-export function isNotInfinitePositiveNumber(v: number): boolean {
-	return v > 0 && Number.isFinite(v);
-}
 
 @component()
 export default class bSlider<T extends object = Dictionary> extends iData<T> {
@@ -62,7 +48,7 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 	 *   *) scroll - scroll implementation
 	 *   *) slider - slider implementation (impossible to skip slides)
 	 */
-	@prop({type: String, validator: (v) => Boolean(sliderModes[v])})
+	@prop({type: String, validator: (v) => sliderModes.hasOwnProperty(v)})
 	readonly mode: Mode = 'slider';
 
 	/**
@@ -80,43 +66,43 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 	/**
 	 * Slide alignment type
 	 */
-	@prop({type: String, validator: (v) => Boolean(alignTypes[v])})
+	@prop({type: String, validator: (v) => alignTypes.hasOwnProperty(v)})
 	readonly align: AlignType = 'center';
 
 	/**
 	 * How much does the shift along the X axis correspond to a finger movement
 	 */
-	@prop({type: Number, validator: isBetweenZeroAndOne})
+	@prop({type: Number, validator: (v: number) => v.isPositiveBetweenZeroAndOne()})
 	readonly deltaX: number = 0.9;
 
 	/**
 	 * The minimum required percentage to scroll the slider to an another slide
 	 */
-	@prop({type: Number, validator: isBetweenZeroAndOne})
+	@prop({type: Number, validator: (v: number) => v.isPositiveBetweenZeroAndOne()})
 	readonly threshold: number = 0.3;
 
 	/**
 	 * The minimum required percentage for the scroll slider to an another slide in fast motion on the slider
 	 */
-	@prop({type: Number, validator: isBetweenZeroAndOne})
+	@prop({type: Number, validator: (v: number) => v.isPositiveBetweenZeroAndOne()})
 	readonly fastSwipeThreshold: number = 0.05;
 
 	/**
 	 * Time (in milliseconds) after which we can assume that there was a quick swipe
 	 */
-	@prop({type: Number, validator: isNotInfinitePositiveNumber})
+	@prop({type: Number, validator: (v: number) => v.isNatural()})
 	readonly fastSwipeDelay: number = (0.3).seconds();
 
 	/**
 	 * The minimum displacement threshold along the X axis at which the slider will be considered to be used (in px)
 	 */
-	@prop({type: Number, validator: isNotInfinitePositiveNumber})
+	@prop({type: Number, validator: (v: number) => v.isNatural()})
 	readonly swipeToleranceX: number = 10;
 
 	/**
 	 * The minimum Y offset threshold at which the slider will be considered to be used (in px)
 	 */
-	@prop({type: Number, validator: isNotInfinitePositiveNumber})
+	@prop({type: Number, validator: (v: number) => v.isNatural()})
 	readonly swipeToleranceY: number = 50;
 
 	/**
@@ -128,20 +114,32 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 	/**
 	 * Initial component options
 	 */
-	@prop({type: Array, required: false})
-	readonly optionsProp?: unknown[];
+	@prop(Array)
+	readonly optionsProp?: unknown[] = [];
+
+	/**
+	 * Factory for an options iterator
+	 */
+	@prop({type: Function, required: false})
+	optionsIterator?: OptionsIterator;
 
 	/**
 	 * Component options
 	 */
 	@field((o) => o.sync.link())
-	options?: unknown[];
+	options!: unknown[];
 
 	/**
 	 * Option component
 	 */
 	@prop({type: String, required: false})
 	readonly option?: string;
+
+	/**
+	 * Option unique key (for v-for)
+	 */
+	@prop({type: String, required: false})
+	readonly optionKey?: string;
 
 	/**
 	 * Option component props
@@ -154,6 +152,23 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 	 */
 	@system()
 	length: number = 0;
+
+	/**
+	 * Link to a content node
+	 */
+	@p({cache: false})
+	get content(): CanUndef<HTMLElement> {
+		return this.$refs.content;
+	}
+
+	/**
+	 * Number of DOM nodes within a content block
+	 */
+	@p({cache: false})
+	get contentLength(): number {
+		const l = this.content;
+		return l ? l.children.length : 0;
+	}
 
 	/**
 	 * Pointer to the current slide
@@ -214,16 +229,14 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 		return 0;
 	}
 
-	/**
-	 * @see current
-	 */
+	/** @see current */
 	@system()
 	protected currentStore: number = 0;
 
 	/** @override */
 	protected readonly $refs!: {
 		view?: HTMLElement;
-		wrapper?: HTMLElement;
+		content?: HTMLElement;
 	};
 
 	/**
@@ -294,16 +307,17 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 	 * Synchronizes the slider state
 	 */
 	@hook('mounted')
-	syncState(): void {
+	@wait('loading')
+	syncState(): CanPromise<void> {
 		const
-			{view, wrapper} = this.$refs;
+			{view, content} = this.$refs;
 
-		if (!view || !wrapper || !this.isSlider) {
+		if (!view || !content || !this.isSlider) {
 			return;
 		}
 
 		const
-			{children} = wrapper;
+			{children} = content;
 
 		this.viewRect = view.getBoundingClientRect();
 		this.length = children.length;
@@ -331,9 +345,9 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 		}
 
 		const
-			{wrapper} = this.$refs;
+			{content} = this;
 
-		if (!wrapper) {
+		if (!content) {
 			return;
 		}
 
@@ -341,7 +355,7 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 			await this.async.sleep(50, {label: $$.syncStateAsync, join: true});
 
 			this.syncState();
-			wrapper.style.setProperty('--offset', `${this.currentOffset}px`);
+			content.style.setProperty('--offset', `${this.currentOffset}px`);
 
 		} catch {}
 	}
@@ -354,10 +368,9 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 	 */
 	async slideTo(index: number, animate: boolean = false): Promise<boolean> {
 		const
-			{length, current} = this,
-			{wrapper} = this.$refs;
+			{length, current, content} = this;
 
-		if (current === index || !wrapper) {
+		if (current === index || !content) {
 			return false;
 		}
 
@@ -369,7 +382,7 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 			}
 
 			this.syncState();
-			wrapper.style.setProperty('--offset', `${this.currentOffset}px`);
+			content.style.setProperty('--offset', `${this.currentOffset}px`);
 
 			return true;
 		}
@@ -386,13 +399,10 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 			current = this.current;
 
 		const
-			{length, $refs} = this;
+			{length, content} = this;
 
 		if (dir < 0 && current > 0 || dir > 0 && current < length - 1 || this.circular) {
-			const
-				{wrapper} = $refs;
-
-			if (!wrapper) {
+			if (!content) {
 				return false;
 			}
 
@@ -406,7 +416,7 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 			}
 
 			this.current = current;
-			wrapper.style.setProperty('--offset', `${this.currentOffset}px`);
+			content.style.setProperty('--offset', `${this.currentOffset}px`);
 			return true;
 		}
 
@@ -453,14 +463,14 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 	@hook('mounted')
 	protected initObservers(): void {
 		const
-			{observers, $refs: {wrapper}} = this;
+			{observers, content} = this;
 
-		if (!observers.mutation && wrapper) {
+		if (!observers.mutation && content) {
 			observers.mutation = new MutationObserver(() => {
 				this.emit('updateState');
 			});
 
-			observers.mutation.observe(wrapper, {
+			observers.mutation.observe(content, {
 				childList: true
 			});
 
@@ -488,9 +498,9 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 		const
 			touch = e.touches[0],
 			{clientX, clientY} = touch,
-			{wrapper} = this.$refs;
+			{content} = this;
 
-		if (!wrapper) {
+		if (!content) {
 			return;
 		}
 
@@ -515,8 +525,7 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 		}
 
 		const
-			{startX, startY} = this,
-			{wrapper} = this.$refs;
+			{startX, startY, content} = this;
 
 		const
 			touch = e.touches[0],
@@ -526,7 +535,7 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 			this.isTolerancePassed ||
 			Math.abs(startX - touch.clientX) > this.swipeToleranceX && Math.abs(startY - touch.clientY) < this.swipeToleranceY;
 
-		if (!wrapper || !isTolerancePassed) {
+		if (!content || !isTolerancePassed) {
 			return;
 		}
 
@@ -541,7 +550,7 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 		this.isTolerancePassed = true;
 		this.diffX = diffX;
 
-		wrapper.style.setProperty('--transform', `${this.diffX * this.deltaX}px`);
+		content.style.setProperty('--transform', `${this.diffX * this.deltaX}px`);
 	}
 
 	/**
@@ -565,13 +574,13 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 		} = this;
 
 		const
-			{wrapper} = this.$refs,
+			{content} = this,
 			dir = <SlideDirection>Math.sign(diffX);
 
 		let
 			isSwiped = false;
 
-		if (!wrapper || !slideRects || !viewRect) {
+		if (!content || !slideRects || !viewRect) {
 			return;
 		}
 
@@ -586,7 +595,7 @@ export default class bSlider<T extends object = Dictionary> extends iData<T> {
 		}
 
 		this.diffX = 0;
-		wrapper.style.setProperty('--transform', '0px');
+		content.style.setProperty('--transform', '0px');
 		this.removeMod('swipe', true);
 
 		this.emit('swipeEnd', dir, isSwiped);