+++ src/super/i-data/i-data.ts
@@ -204,6 +204,10 @@ export default abstract class iData<T extends object = Dictionary> extends iMess
 
 	/** @override */
 	initLoad(data?: unknown, silent?: boolean): CanPromise<void> {
+		if (!this.isActivated) {
+			return;
+		}
+
 		const label = <AsyncOpts>{
 			label: $$.initLoad,
 			join: 'replace'
@@ -840,7 +844,7 @@ export default abstract class iData<T extends object = Dictionary> extends iMess
 	 */
 	protected onDelData(data: unknown): void {
 		if (data != null) {
-			this.db = undefined;
+			this.db = this.convertDataToDB(data);
 
 		} else {
 			this.reload().catch(stderr);